convert_command

+++ app/models/security_camera.rb
@@ -63,12 +63,12 @@ class SecurityCamera < ApplicationRecord
                 final_path="#{self.camera_image_dir}/#{DateTime.now.to_i.to_s}.jpg"
                 convert_command="convert #{tmp_path} -resize 1920x1080^ -gravity center -quality 75 #{final_path}"
                 puts "[convert] #{convert_command}"
-                if system(convert_command)
+                if system(convert_command)&&File.file?(final_path)
+                    puts "[convert] ok"
                     self.update_attribute(:last_seen,DateTime.now)
                     FileUtils.rm(tmp_path)
                     REDIS.del("screenshot:timer")
                     self.update_attribute(:error_count,0)  
-        
                     return true
                 else
                     self.update_attribute(:error_count,self.error_count+1)
